feat(rust): add specific view for `OpenUrl`
